indexpack: Add a VFS for read-only access to a ZIP file.
This change makes it possible to open a zip file containing in index
pack directory, and read out compilations and files without having to
first unpack the file.  This variant is read-only, and the operations
that write will report errors.  We may change that in the future, but
for now we only need read.