Rework the graphstore interface from channels to callbacks.
The main effect of this change is to allow graphstore operations to be
cancelled without extra channels.  Graph store operations that produce entries
now take a callback, and if the callback returns an error, the operation is
terminated early.  io.EOF terminates the operation without error; all other
errors are propagated back to the caller.

This change also lets us drop a bunch of helper functions that were doing more
or less the same thing.