Reproduce header search state and use a VFS for (k)index( packs).
This CL addresses file lookup from index packs in a more thorough
way. Previously we used Clang tooling support for virtual files
and attempted to canonicalze paths using the working directory
at the time of extraction. This was not a complete solution, as
some paths with relative elements (foo/../bar) would not be
correctly handled. In fact, one root cause of T24 was that
relative paths were not mapped to virtual files, but instead were
looked up in the real filesystem. By using a virtual filesystem,
we can prevent (most) access to the real filesystem, thus catching
these kinds of errors in the future.

The other root cause of T24 is that the extractor can set up
a header search state that is not reproduced in the indexer.
Absent any manipulation, the header search state is a function
of the machine on which Clang is currently running. This leads to
confusion when the indexer is run on a machine distinct from the one
that ran the extractor. Unfortunately, Clang's representation of
the search state depends on objects (DirectoryEntry*s) that cannot
be reasonably constructed without control of the VFS layer.

This CL does make the assumption that paths are tree-structured
(/usr/lib/gcc/x86_64-linux-gnu/4.9/../../../../include/c++/4.9/string
is the same as /usr/include/c++/4.9/string). Should this be a problem,
we can complicate path canonicalization, perhaps by checking in the
extractor whether relative paths have this property or not and
blacklisting the ones that fail.

Tests were changed minimally. When not reading from a kindex
or index pack, the indexer will still access the local filesystem.