Sample web ui: store state in url and restore during page load
Page state currently consists of the currently viewed file's VName as well as
the most recently jumped to line/offset.  State can be manually constructed
since the VName is considered partial and used in a search query.  Thus, only a
file's path can be given, and if its unique, the correct file will be navigated
to without needing to know its signature, language, corpus, or root (although,
of course, these can be provided to narrow the search).

Example URLs:
  /#kythe/java/com/google/devtools/kythe/platform/indexpack/Archive.java
  /#kythe/java/com/google/devtools/kythe/platform/indexpack/Archive.java?corpus=kythe
  /#kythe/java/com/google/devtools/kythe/platform/indexpack/Archive.java?line=62
  /#?signature=782003c74726ae139fe7223819543426dd812d68b7f5655bf41af0e75c875612
  /#kythe/java/com/google/devtools/kythe/platform/indexpack/Archive.java?language=java&corpus=kythe&signature=782003c74726ae139fe7223819543426dd812d68b7f5655bf41af0e75c875612&line=300