Prepare to remove prebuilt LLVM binaries.
Summary:
This CL is the first in a series that will remove
prebuilt LLVM binaries and require developers to
point campfire toward LLVM on their own. It introduces
three new properties to .campfire_settings:

llvm_repo, a path to an LLVM checkout, used to make sure
  that the version is up to date
llvm_lib, a path (relative to third_party/llvm) to use
  to look for built LLVM libraries
llvm_include, a path or array of paths (relative to the
  build root) to add to the include search list.

buildtools/check_modules.sh uses the llvm_repo
property to interrogate LLVM checkouts. It supports
git or svn. In this CL, the assertions about minimum
checkout versions are disabled. This script is run
each time campfire is invoked.

I have also amended some extractor tests to make them
more robust in the face of changing compiler resource
files.

Reviewers: schroederc, jdennett, mazurak, fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D128