Require LLVM to be built from source
Summary:
This CL configures the build system to expect
LLVM to be built from source.
It enables the check in ./campfire that verifies
that a local LLVM checkout exists and is up
to date. If this check fails, update_modules.sh
knows how to pull down LLVM and build it in the
default location. (update_modules.sh --docker will
clone LLVM on the host and build it inside Docker
so that campfire-docker still works.)

Removing the old binaries and headers will be done
outside of phab owing to technical limitations.

Reviewers: jdennett, schroederc, mazurak, fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D135