Skeleton for a tool that uses Kythe to find missing includes.
Summary:
fyi is a tool that adds missing #includes to a
source file. It does this by running multiple
compile passes, first to collect information about
undefined symbols, then later to check to see if
the includes it adds have fixed discovered problems.

This CL merely initializes Clang tooling support
and reparses a source file a few times. You can see
when precompiled headers kick in based on the parse
timing (there's an internal counter in the compiler
that waits two parses and builds a PCH on the third,
so you'll see something like [3.5s, 3.5s, 4.8s, 1.1s,
1.1s]).

(Expect to see "Error while processing .../fyi/testdata/nothing_test/nothing.cc."
when you run the test; this happens because the skeleton only knows how
to break files right now. The test checks that the tool yields nonzero
but that it emits the correct modified source.)

Note also that for nontrivial programs PCH is broken
until D135 lands (it's an LLVM bug).

Reviewers: jdennett, mazurak, schroederc, fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D136