C++ indexer: index base classes.
Summary:
This diff emits extends/ edges for C++.

There are a few TODOs remaining about
linking to particular template instantiations
instead of `tapp` specifications of these
(and one about T27). This will be better
to deal with in a separate CL, as I think we'll
want to do it consistently inside
BuildNodeIdForType.

Reviewers: schroederc, jdennett

Reviewed By: jdennett

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D144