Fix a small issue with campfire/arcanist support
Summary:
Without this redirection to /dev/null and in a clean checkout of the
repository, the campfire root directory is picked up as "target" to be
unit tested by arcanist.  This causes `arc unit` to fail incorrectly.

This change also adds some more files to the blacklist for copyright
checks.

Reviewers: fromberger, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D172