C++ proto macro
Summary:
We can't write a protoc rule, but we can write a macro
that generates cc_libraries like //kythe/proto:any_proto_cc.
Existing targets (//kythe/proto:any_proto) are the same as
before. The macro is named genproto_all because the rule
used to generate the Java proto libraries must be named
exactly "genproto". Targets must be named "foo_proto" and
they must be in a package equivalent to their path from
the workspace root.

Reviewers: schroederc

Reviewed By: schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D174