Add critical edges (that resist removal by claiming); deal with embedded decls.
Summary: Because static claiming only takes preprocessor effects into account, it's not able to distinguish between interesting cases that only come up after typechecking. These include DeclRefExprs, as these depend on the binding context of the program, which is not available during preprocessing. We call an edge in the graph that claiming cannot account for *critical*. This CL adds the facility to label certain edges as critical.

Reviewers: jdennett, fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D222