Add BazelTestEngine for arcanist
Summary:
This change only adds the engine.  It doesn't make arcanist use it yet.
To do so, replace "CampfireTestEngine" in .arcconfig with
"BazelTestEngine".

The main difference between the engines is that the new engine does not
list non-test targets in the test results.  They are, however, built
during testing and a failure to build them will throw an exception.

Reviewers: zarko

Reviewed By: zarko

Maniphest Tasks: T29

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D250