OSX fixes: 6l -s, tell asciidoc which Python to use and where dot is at.
Summary:
On OS X, asciidoc fails to find the executable of the
Python interpreter, leading it to try and execute the
empty string. We can fix this by overriding the
autodetected interpreter with an attribute.

The environment Bazel runs builds in does not include
the path for graphviz on OS X. This CL adds logic to
configure this in setup_bazel.sh.

Finally, there's a Go issue on OS X that we're hitting.
(https://github.com/golang/go/issues/10254). Stop hitting it.

Reviewers: shahms, fromberger, schroederc

Reviewed By: schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D319