Add flow and b64_ocaml.
Summary:
This diff adds flow and b64_ocaml to third_party,
as well as stubs for a JavaScript indexer. Rather
than try to call into Flow, we instead initially plan to
add a new Flow command (dump-kythe) that makes use
of the existing client/server framework. Adding this
command requires applying a small patch to Flow proper
to register the client- and server-side code; we also
add symlinks to the Kythe stubs to make use of Flow's
existing ocamlbuild-based build.

This will be an optional feature. We will not
require Kythe users to install ocaml to build or
test Kythe.

Reviewers: fromberger, jdennett

Reviewed By: fromberger, jdennett

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D362