indexpack: ReadUnits passes the digest to the callback with the unit.
Summary:
The ReadUnits method of an *indexpack.Archive can be used to scan the contents
of the pack, but there is currently no good way for the callback to learn the
digest of the unit it's given.  This change adds an extra argument that passes
the digest string.

Tests and consumers fixed up.

Reviewers: shahms, schroederc

Reviewed By: shahms, schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D440