Deduplicate output early using a memcache.
Summary:
This diff gathers output entries into bunches based on
higher-level semantic signals. Before being emitted, these
bunches are hashed and the hash is checked against a memcache.
If the memcache holds the hash, then the bunch is dropped on
the floor. Bunches are merged together to maintain a
minimum size (to avoid overtaxing the memcache) and they are
split to avoid a maximum size (to make hashing more effective).

With the cache turned off, the speedup for raw computation
is ~0.97 (e.g., there is a slight slowdown) for the previous
set of four benchmark units. With the cache off and when we're
writing to a graphstore with the *cxx_common* .kindex files
that we used to benchmark the hash patch (D450), the speedup is
~1.04. With the cache on, the speedup for the *cxx_common*
files is ~1.39. (The baseline for this speedup is the time
taken for *cxx_common* with vname hashing patched in; without it,
the speedup from pre-D450 is ~1.65.)

Reviewers: fromberger, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D452