Experimental xrefs.CrossReferences API
Summary:
This change adds a new API call to the xrefs service (see xref.proto) and an associate sub-command to the kythe CLI (see `kythe help xrefs`).

NOTE: this call is **currently** implemented using Nodes/Edges calls and does not have an associated serving table.  This means that it can be relatively slow compared to the other APIs.

Reviewers: jdennett, fromberger

Reviewed By: fromberger

Subscribers: amshali

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D504