Specify edge ordering in xrefs/assemble for edges serving data
Summary:
The new ordering helps ensure that pertinent edges (anchor edges and
primary Kythe edges for names, types, etc.) are usually on the first
EdgesReply page.  As a side-effect, this produces better ordering
behavior for the experimental CrossReferences API call.

This change also defers the ordering of the xrefs in the sample web UI
to the order in the EdgesReply.

Reviewers: zarko, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D511