Go extractor: Don't capture non-package test files.
Summary:
When go/build reads a package, it lists test files that do not declare
themselves as part of the package under test.  We do not want to include
these in the compilation for that package, as it is illegal to combine
files declaring different packages in one compile step.

In the long run, we should extract out-package test files as separate
compilations with dependencies on the original; however that will
require some other changes to the extractor, so for now I'm just
omitting the out-package files from the compilation to keep the
type-checker happy.

Ref. T62.

Reviewers: schroederc

Reviewed By: schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D535