Rebase LLVM.
Summary:
Update LLVM to pull in new support (target-sensitivity for the extractor)
and so that we don't fall too far behind.

Test Plan:
Some hashes for Darwin (OS X) need to be updated in a subsequent CL.
There's also been a change re: finding the standard library on OS X
that I need to figure out.

Reviewers: fromberger, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D568