Add an experimental claim client.
Summary:
This CL adds support for dynamic claiming of transcripts
by using a memcache instance. When this option is turned
on with an experimental flag, many more output entries are allowed
to be eliminated. This will result in an incomplete graph. At this
point you almost certainly do not want to enable the option.

Reviewers: schroederc, shahms, fromberger

Reviewed By: fromberger

Maniphest Tasks: T67

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D564