Go: A library to type-check a Go compilation unit.
Summary:
Package indexer resolves dependencies and runs the type checker, as a
prerequisite to generate Kythe cross-references.

At present, there is one limitation: In order to type-check compiled Go
packages, we need to read the type headers from the .a files generated
by the gc compiler.  But the gcimporter package in the standard library
was moved to an "internal" directory, which means we can't build this
package with 'go build' because of the visibility restriction.

The go/importer package has a hook that would allow us to avoid direct
use of the gcimporter package, but the hook is not fully implemented.
So as a temporary workaround, this commit depends directly on the
gcimporter package, and I'll look separately into implementing the hooks
as an upstream contribution for Go 1.6.

This change includes unit tests for the Resolve function.  More extensive tests
will be added with the upcoming commits for the backend (since that's where the
vnames are used).

Reviewers: shahms

Reviewed By: shahms

Maniphest Tasks: T32

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D537