Populate denormalized serving table format
Summary: This change changes `write_tables` to populate the denormalized format introduced in D577.  This then allows for the calls to `xrefs.Nodes` in the implementations of `xrefs.Edges` and `xrefs.Decorations` to be removed.  Instead, the facts for the source/target nodes of edges and the facts for the target of each file decoration are retrieved with each edge and file decoration, respectively.

Reviewers: shahms, zarko

Reviewed By: zarko

Subscribers: amshali

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D580