Add CrossReference message to serving.proto
Summary:
This message will be used in future changes as an intermediate structure
for building the pre-cached responses to the Decorations and
CrossReferences methods.

Alongside this new message, the existing FileDecorations message has had
its Anchor message moved to the top-level and
the (file_ticket, source_text, encoding) tuple has been moved into its
own message.

Note: this is a breaking change to the serving tables format, but this
change is already bundled with quite a few other breaking changes to the
format since the last release.

Reviewers: shahms, fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D626