Serving pipeline: replace LevelDB sorts with simple external sorter
Summary:
The intermediary stages for the partials edges, complete edges, and
decoration fragments only need to be sorted and then read sequentially.
Previously, this was done, for simplicity, with the existing LevelDB
machinery (carefully encoding each entry's key to get the correct
sorting), but it turns out that using a general key-value store for this
purpose is actually kind of slow.

This change introduces a new package, disksort, which was custom built
to take over LevelDB's job with less overhead.  Each element is now kept
in-memory (up to a specified number of elements) and then a set of data
is paged to disk as a sorted, delimited stream of bytestrings.  When all
elements have been added, the sorted shards are merged back together
incrementally and given to the user sequentially.

Quick Performance Numbers (write_tables --graphstore gs --out tbl):
  * 188,798 entries in GraphStore
    - Before: 7.20s
    - After:  5.85s

  * 13,970,489 entries in GraphStore
    - Before: 1562s
    - After:   548s

Reviewers: zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D631