Serving pipeline: write PagedCrossReferences to the resulting serving table
Summary:
The PagedCrossReferences structure is similiar to a PagedEdgeSet with a
top-level message used as the root for the CrossReferences/Edges lookup
and zero-or-mode Pages stored separately to reduce the size of a single
lookup.  The pipeline constructs the new data using the same SetPager as
the EdgeSetBuilder.

This change only contains the logic to write cross-references data to
the serving table.  A separate change will have the logic to read/serve
the data.

Reviewers: shahms, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D641