write_tables: add --compress_shards flag to reduce size of intermediate data
Summary:
Add CompressShards option to disksort package that determines whether
each temporary file shard is snappy compressed.  Compression is optional
since it adds ~15% of overhead to write_tables.

This change also allows for the size of the IO buffers in disksort to be
configured and increases their default size from 4096 to (2<<13).

Reviewers: shahms, fromberger, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D647