Allow C++ tests to fail indexing xor verification
Summary:
It's useful to be able to check in failing tests to document
unimplemented features.

This CL amends the verifier to return status code 2 when
one of its input scripts is malformed (for example, it may contain
an assertion syntax error or a dangling @-reference).

This CL also cleans up some of the C++ test scripts.

Reviewers: fromberger, shahms

Reviewed By: shahms

Subscribers: jdennett

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D668