indexpack: Use a sensible default UnitType.
Summary:
If an index pack is created or opened with no UnitType option, it will
panic when trying to read compilations out.  Instead, make the default
be json.RawMessage, which is an alias for []byte.  This will permit a
reader to view the contents of the pack without choosing a decoding in
advance.

Tests and documentation updated.

Reviewers: schroederc, shahms, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D674