Distinguish decl/def callables and don't prefer first decls
Summary:
The previous mechanism for building callable IDs
was insufficiently resistant to ODR-style problems.
Previously we tried to assign the def and decl of
a function the same callable, e.g. all of:
  void f();
  void f();
  void f() { }
would be callableas the same node. This fails to
resist ODRish problems at the expense of extra
complication (in consing up a bespoke callable ID)
because only the information common to all of the
decls could be considered--so the only ID we could
decide on would be equivalent to "void ::f()".

This diff changes ID generation to derive a callable
signature from the decl it's %callableas. We can link
defns with decls using completes edges (the function
of which was unnecessarily duplicated by the old
callable ID scheme).

This diff also stops preferring the first function
decl for generating callableas/overrides edges, since
these ought to be generated for every appropriate function.
This fixes cases where out-of-class definitions of
member functions get callable nodes that never are given
kind facts.

Reviewers: jdennett, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D677