Add build-tagged exporter wrapper.
Summary:
The gcimporter package we use to read Go package archives is changing
interface with Go 1.6. Unfortunately, the "importer" package we should
be using instead still does not support custom hooks, so we are stuck
with using this for now.

To bridge the migration from Go 1.5 to 1.6, this change adds a wrapper
function for the call, with two build-tagged variants. Bazel does not
understand the build tags, but having both makes it easier to patch the
BUILD rule on import.

Addresses T94.

Reviewers: zarko, schroederc

Reviewed By: schroederc

Projects: #go

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D682