Don't check the CU hash for //kythe/extractors/cmake:compilation_database
Summary:
The VName for a compilation unit is calculated in part from
hashing all of its arguments. The first argument is the path
to the extractor as the extractor was run. In this test, that
path is absolute, so the CU hash is machine-dependent.

We might consider ignoring arguments[0] for computing the CU hash.

Reviewers: schroederc, fromberger, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D721