Make anchor VName conventions mandatory
Summary:
Right now, I believe our indexers all implicitly obey the following rule:

If an anchor is `childof` some file, then that anchor's VName should have the
same `corpus`, `path`, and `root` as that file. Otherwise, at least the `path`
field of the anchor's VName should be empty.

This diff makes this mandatory. This makes it much cheaper to locate an anchor's
source file without incurring the cost of following the childof edge. It also
reduces the amount of denormalized information we need to store to make calls
like CrossReferences efficient.

Reviewers: fromberger, schroederc, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D733