Implement Callers using existing APIs
Summary:
This diff implements the Callers API using existing
APIs (Edges and CrossReferences). It's a first step
toward a version of Callers that performs well enough
to be used in production and should be easy to hack on
while we determine exactly what sort of data we need
to denormalize.

The -icorpus and -normalize_file_vnames flags make it
easier to generate API test data from .cc files and
.kindexes.

Reviewers: fromberger, schroederc

Reviewed By: schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D724