Hash enum constants more efficiently.
Summary:
Profiling on one of LLVM's largest TUs revealed that we were spending a
large fraction of our time hashing enum constant names. I found that this was
because we were re-hashing all the names as part of the process of creating
a node name for every constant, effectively causing O(n^2) complexity. To
reduce that complexity, memoize enum hashes so the result can be re-used for
enum constants. This change roughly halved indexer run time for that TU.

Reviewers: schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D779