Fix apparent logic error in file output stream implementation.
Summary:
We were previously testing whether MergeDownIfTooSmall returned false to
determine whether a buffer was merged. This was wrong; MergeDownIfTooSmall
returns true if a buffer was merged. As a result, we were never moving buffers
onto the free list by popping them off the stack. This change fixes the sense
of the test, resulting in a reduction in memory consumption of about 300%
in one of LLVM's largest TUs.

Reviewers: schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D782