Add and use check_version macro to require bazel release 0.2.0-2016-03-21
When 0.2.1 is released (which should be soon) I plan to update the required
version to that and maintain compatibility with that (or subsequent) releases
going forward.

Summary:
...rd.

The parenthesized commit number isn't actually checked, but is printed on
failure and gives an easy way to get a working version.

Add and use check_version macro to require bazel release 0.2.0-2016-03-21

Reviewers: schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D783