Classify and optionally drop some redundant template-related nodes and edges
Summary:
This CL classifies childof edges that relate anchors to template
instantiations as "childof/context" edges. It also provides an
experimental indexer flag that drops anchors and edges that are
template instantiation independent. For example, in the program:

```
  template<typename T> class C { int i; };
  template class C<short>;
```

the anchor covering `i` will be emitted twice, once in `class C`
and once in `C<short>` with a `childof/context` edge to `C<short>`.
Similarly, the anchor covering `int` and its `ref` edge will be emitted
twice unless the new option is asserted, in which case only those related
to the primary template will be emitted.

Reviewers: fromberger, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D790