Add basic support for running autotools configure to generate files
Summary:
This change adds a configure_generate() macro which will run a package-local
configure script, setting up the environment appropriately.

Additionally, this factors out the common behavior from this and the earlier
cmake support into a separate tool and switches cmake to use it.

Finally, this is intended to be used to externalize the googlelog dependency.

Reviewers: zarko, schroederc, fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D789