Create virtual directories for include search paths.
Summary:
Consider the following scenario:
- a compilation unit adds a directory foo to the include search path via -Ifoo
- that directory's parent contains a header file named foo.h included via
  #include "../foo.h"

In that scenario, Clang will during the indexer phase stat the directory foo
to determine whether to add it to its search path. If we did not extract
any files from that directory, there will be no VFS entry for it and the
stat will fail. Clang will subsequently fail to find "../foo.h" due to the
missing search path entry.

To solve that problem, teach VFS to accept a list of virtual directories,
and have the indexer pass the list of search paths as that list. During
the extraction phase, that list had already been filtered through the same
mechanism to include only existing directories.

Reviewers: schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D781