xrefs service: add ordinal field to each EdgeSet edge
Summary:
Change the kythe.proto.EdgeSet.Group and
kythe.proto.CrossReferencesReply.RelatedNode proto definitions to
include an edge's ordinal (if one exists) alongside its target ticket.
This change does not yet actually surface any ordinals because that
information is not in the serving data.  That will be left for a later change.

NOTE: this is a BREAKING CHANGE to the xrefs API

Reviewers: fromberger, zarko

Reviewed By: zarko

Subscribers: juanch, #core_team

Maniphest Tasks: T114

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D801