Add definition locations for file decorations
Summary:
If the `DecorationsRequest.reference_definitions` switch is `true`, each
reference in the `DecorationsReply` will have an associated definition
location for the reference's target.  The target definition is only
populated if it is unambiguous.  For certain nodes, the definition may
be indirect.  The only such indirection is through `callableas` edges
right now.

NOTE: this is currently experimental and the implementation is slow

Reviewers: zarko, fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D809