Go: Clean up the interface for reading ZIP files.
Summary:
Rather than passing in an io.ReadSeeker and adapting it to the
io.ReaderAt that the archive/zip package requires, make the user pass in
the reader and length explicitly. This improves performance in cases
where the filesystem has a faster implementation of ReadAt than a seek
followed by a read.

Reviewers: zarko, shahms, schroederc

Reviewed By: shahms, schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D813