Index C++ namespaces as first-class nodes.
Summary:
This diff emits package nodes for namespaces. Named
namespaces in headers and source files adopt the corpus of the
file they're declared in (but not the root or path). Anonymous
namespaces in source files or in .inc files that are *not* accessible
from header files* adopt the main source file's root, path, and corpus
(as those anonymous namespaces are specific to that main source
file). Anonymous namespaces that are accessible through header files
are named like named namespaces (in that they adopt the corpus of the
enclosing file).

* For a given compilation--so if a .inc is used inconsistently
across a project, the objects inside may be duplicated.

Reviewers: shahms, fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D810