Refactor the local system dependencies such that they are more correct and platform agnostic.
Summary:
This builds on and cleans up the previous kythe_pkg_config repository rules
such that a local directory can be specified via environment variable
to force {lib/,include/} style dependencies on all platforms.

It also supports pkg-config on all platforms, with the caveat that
OS X Xcode 10.11 has a broken OpenSSL installation. It includes the libraries
and .pc (pkg-config) file, but no headers.

Installing and linking with --force via homebrew works, but it's almost
as if Apple doesn't want people developing on/for their platform.

Reviewers: schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D835