Add experimental formats for C++.
Summary:
It's expensive to store types and signatures (which contain types)
as literal strings. Formats compress types and signatures by
defining them recursively. For example, the format for a
C++ pointer is `%1.*`. The format for a variable `foo` is
`%^::foo`. The format for an arbitrary type application is
`%0.<%1,>`.

This feature is explicitly marked experimental as I expect the
definitions of format tags and the formatting algorithm to
change as we discuss and apply it.

This addresses T26.

Reviewers: fromberger, shahms

Reviewed By: shahms

Maniphest Tasks: T26

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D843