Java indexer: eliminate possible AssertionError
Summary:
The normal behavior of `TypeTag#getPrimitivTypeKind()` (and
`JCPrimitiveTypeTree#getPrimitiveTypeKind()` by proxy) throws an
exception if the tag is an `ERROR` type.  This bypasses the exception to
instead just treat errors as normal types so that we can continue
analyzing the rest of the compilation.

This change also adds a `--verbose` flag to the indexer to emit debug
messages.

Reviewers: zarko, shahms

Reviewed By: shahms

Subscribers: juanch

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D878