Fix go rules wrt native libraries
Summary:
cgo directives in libraries will add "-lfoo" to ldflags when looking for a
libfoo.a; the compiler driver gets upset if it can't find libfoo.a using a -L
path even if it's also provided with an absolute path to libfoo.a. To fix this,
we provide both (since providing just the -L versions won't work as some
libraries like levigo have special rules regarding linking in other dependencies
like snappy).

Reviewers: fromberger, schroederc

Reviewed By: schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D893