Begin merging Callers API into CrossReferences
Summary:
This diff changes xref.proto to look as though Callers has been merged
into CrossReferences. In particular, it:
  - makes Printables top-level messages
  - replaces Anchors in CrossReferenceReply with RelatedAnchors,
    which contain Anchors, Printable display names, and a list
    of specific sites (also Anchors)
  - adds a CallerKind enum to CrossReferencesRequest to specify
    what kind of callgraph information to return.

The merged API lacks some fields that are currently in Callers (but
that are not currently hooked up to anything). Should these be necessary,
we can always add them to RelatedAnchors at a later time.

There should be no functional changes.

Reviewers: schroederc, fromberger, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D914