Initial Hello World for rust compilation by bazel
Summary:
Prepping for the implementation of the rust indexer,
this commit sets up bazel builds for rust. Currently it only
builds a hello world program but the project is split into
two different components a library and an executable.

Test Plan:
Run bazel build //kythe/rust/indexer:release. If rust is
present it will build, and //kythe/rust/indexer:test will run the
tests. If rust is not present both will fail and display an error
message.

Reviewers: zarko, schroederc

Reviewed By: schroederc

Projects: #rust

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D931