Fold Callers into CrossReferences
Summary:
This diff adds callgraph support to (serving table-based)
cross-references. It is an intermediate step toward full
support as it still defers to the Slow* functions. The diff
also removes the old `callers` subcommand from the Kythe
command-line tool. Requesting callgraph information from the
`xrefs` subcommand is off by default.

The diff modifies CrossReferences API such that References
to callgraph-related nodes can be excluded or reported
exclusively. The former is useful because it reduces duplication
when asking for both references and callgraph data; the latter
is chiefly used internally.

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D946