Emit refs for locals in Rust
Summary: By iterating over the def_map we can find all references to something defined elsewhere. There are currently some refs that are missed because of macro expansion, but to work around them we'd need to use a secondary lint pass. Currently we're marking refs and definitions during the check_crate_post which runs after macro expansion. This allows us to ensure it runs after check_local which is where our blacklist is built. A multi-pass approach should be covered in a different CL though.

Test Plan: Added a ref test

Reviewers: shahms, schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D968