Fix crash on indexing callgraph for impls
Summary: When indexing callgraphs we inspect the parent item to determine it's something we want to attach a childof edge to. Unfortunately, expect_item was causing a crash when inspecting parents which were method impls because they aren't of type Item. The new method safely inspects the parent by only acting on the call if it is of one of the two appropriate types.

Test Plan: Added a test for impl_methods

Reviewers: zarko, schroederc, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D980