xref.CrossReferences: hide SlowSignature implementation behind feature flag
Summary:
Since the signature generation aspect of the CrossReferences method can
cause latency issues, hide it behind a feature flag that is disabled by
default.  Once the latency issues have been addressed, the flag will be
removed and the feature will always be enabled.

The guard to SlowSignature in SlowCallersForCrossReferences is mostly to
be consistent.  The "slow" part of SlowCallersForCrossReferences will
still be true.

Reviewers: zarko

Reviewed By: zarko

Maniphest Tasks: T156

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D988