Fix a null dereference while indexing types
Summary:
It's possible (nay, plausible) that the deduced form
of a type (DT/DType) has a different tag than the TypeLoc-
identified form of a type. Elsewhere in BuildNodeIdForType,
we guard against this by always checking if dyn_casts on
DTs to the TypeLoc tag are null (and if they are, we use
the TypeLoc type instead). This was missed in handling
DependentSizedArrays, so in some cases we may try to call
getSizeExpr() on null. This diff fixes that bug.

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D991