Update linter.sh to properly test for jq
Summary:
The script previously attempts to use one invocation of `command` to test for both `shellcheck` and `jq`; while this would guard against shellcheck not being present, it would allow the block to be executed even if `jq` is not present.

Use two separate `command` invocations instead, and only execute the block if both commands succeed.

Test Plan:
- install shellcheck
- uninstall jq
- `arc lint tools/arc/linter.sh`

Reviewers: schroederc

Reviewed By: schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D992