Add partial support for Objective-C.
Summary:
Add support for methods and interfaces. This does not include categories or
protocols. They require more work to fully represent in the graph. Instance
variables are also not supported yet. The todos in the code list many other
situations that are not yet fully implemented that are related to interfaces
or methods.

Todos are present for langauge features that require further investigation.

Test Plan:
A collection of verfication tests under testdata/objc. Some tests
are currently disabled because the feature they test is not yet implemented.

Reviewers: zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1001